Create README.md
# Zuri--CRUD-App-With-Database

# CRUD application with database


This is a simple nodeJS application that connects to a database(mongoDB) and performs the CRUD operations. It is a task on CRUD relationship with database, assigned by the Zuri team. 
All routes were tested by Postman app to ensure full functionality.


## Acknowledgements

 - [The Zuri Team](https://training.zuri.team/)
 - [Ingressive For Good](https://ingressive.org/)

  
## Author

- [Tominiyi Orisatuyi]

Host link?: Please bear with me, I'm having power supply issue at the moment. it should be ready in 2 hours from now. thank you